bank account

+++ programming_paradigm/robust_division_calculator.py
@@ -1,6 +1,6 @@
 def safe_divide (numerator,denominator ):
   try:
-      result = numerator / denominator
+      result = float(numerator / denominator)
       if denominator == 0:
         raise ZeroDivisionError ("Error: Cannot divide by zero.")
       else:
